// src/lib/schemas/auth-schemas.ts
import * as z from 'zod'

export const signupSchema = z.object({
    name: z.string().min(2, 'Name must be at least 2 characters.'),
    email: z.email('Please enter a valid email address.'),
    password: z.string().min(8, 'Password must be at least 8 characters.'),
})

export const loginSchema = z.object({
    email: z.email('Please enter a valid email address.'),
    password: z.string().min(8, 'Password must be at least 8 characters.'),
})